+++ test/DotnetPackaging.Tests/Deb/EndToEnd/ExecutableMetadata.cs
@@ -1,0 +1,5 @@
+﻿using DotnetPackaging.Deb;
+
+namespace DotnetPackaging.Tests.Deb.EndToEnd;
+
+public record ExecutableMetadata(string CommandName, DesktopEntry DesktopEntry);
