+++ test/DotnetPackaging.Tests/Tar/TestMixin.cs
@@ -1,4 +1,5 @@
 ﻿using System.Reactive.Linq;
+using Zafiro.FileSystem;
 
 namespace DotnetPackaging.Tests.Tar;
 
@@ -14,5 +15,11 @@ public static class TestMixin
     {
         await using var stream = File.Create(path);
         await stream.WriteAsync(bytes.ToEnumerable().ToArray());
+    }
+
+    public static async Task DumpTo(this IObservable<byte> bytes, IZafiroFile file)
+    {
+        await using var stream = new MemoryStream(bytes.ToEnumerable().ToArray());
+        await file.SetContents(stream);
     }
 }
